feat(api): Move SDK schemas to @packages/database
