+++ packages/api/package.json
@@ -20,7 +20,6 @@
       "typecheck": "tsc --noEmit --emitDeclarationOnly false"
    },
    "dependencies": {
-      "@contentagen/sdk": "workspace:*",
       "@packages/workers": "workspace:*",
       "@packages/authentication": "workspace:*",
       "@packages/database": "workspace:*",
